% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{algorithm}
\alias{algorithm}
\title{Optimization algorithm}
\arguments{
\item{fn}{The function to be optimized, taking a numeric vector of parameters
and returning a numeric value.}

\item{lower}{A numeric vector, lower boundary of the parameter vector.}

\item{upper}{A numeric vector, upper boundary of the parameter vector.}

\item{control}{A list of further control parameters for the algorithm.}
}
\value{
A list containing \code{par}, the parameter combination of the optimum,
\code{value}, the optimal function value, and \code{trace}, a \code{data.frame} of parameter
values visited during the algorithm's run.
}
\description{
Optimization algorithm
}
