% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simann.R
\name{genptry}
\alias{genptry}
\title{Generate a candidate point in the neighborhood}
\usage{
genptry(p, scale)
}
\arguments{
\item{p}{a numeric vector}

\item{scale}{a real number, the radius of the Gaussian}
}
\value{
a numeric vector
}
\description{
Using a Gaussian kernel of standard deviation \emph{scale} around the current
point, this function generates a new point nearby.
}
