% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_fn_name}
\alias{get_fn_name}
\title{Get a name for the function value}
\usage{
get_fn_name(x)
}
\arguments{
\item{x}{a named object}
}
\value{
a character, either name of \code{x} or \code{"fn"}
}
\description{
Get a name for the function value
}
