% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_par_names}
\alias{get_par_names}
\title{Get names for the parameter vector}
\usage{
get_par_names(x)
}
\arguments{
\item{x}{a named object}
}
\value{
a character vector, either the names of \code{x} or strings of the form
\code{"par1"}, \code{"par2"}, etc.
}
\description{
Get names for the parameter vector
}
