% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch.R
\name{gridsearch}
\alias{gridsearch}
\title{Grid Search Algorithm}
\usage{
gridsearch(
  fn,
  lower = NULL,
  upper = NULL,
  step = NULL,
  axes = mapply(seq, from = lower, to = upper, by = step, SIMPLIFY = FALSE),
  grid = expand.grid(axes),
  control = NULL
)
}
\arguments{
\item{fn}{The function to be optimized, taking a numeric vector of parameters
and returning a numeric value.}

\item{lower}{A numeric vector, lower boundary of the parameter vector.}

\item{upper}{A numeric vector, upper boundary of the parameter vector.}

\item{step}{A numeric vector of step widths for each dimension of the
parameter space.}

\item{axes}{A list of numeric vectors, defining the axis values of the grid
for each dimension of the parameter space.}

\item{grid}{A data frame of all parameter combinations to be visited.}

\item{control}{A list of further control parameters for the algorithm.}
}
\value{
A list containing \code{par}, the parameter combination of the optimum,
\code{value}, the optimal function value, and \code{trace}, a \code{data.frame} of parameter
values visited during the algorithm's run.
}
\description{
The grid search algorithm searches the parameter space of a function \code{fn} for
the optimal parameter values on a pre-defined grid.
}
\details{
The grid can be defined by either
\enumerate{
\item lower and upper bounds of the parameter space together with
a step width for each dimension,
\item axis vectors for each dimension of the
parameter space, or
\item a "grid" containing every parameter combination to be visited.
}
}
\section{Control parameters}{


The list \code{control} may contain the following objects:
\itemize{
\item \code{fnscale}: A numeric, if it is non-negative or NULL, the algorithm searches
for the minimum, if it is negative, it searches for the maximum.
\item \code{REPORT}: An integer, if it is \code{NA_integer_} or negative, no trace is
reported. If \verb{>=0}, a trace is reported. If \verb{>0}, status updates are sent
to a \code{progressr} handler every step. By default, \code{REPORT = 0}.
}
}

\examples{
fn <- function(vec){
  return(c(result =
  (-1)*dnorm(vec[["x"]], mean = 3.9)*dnorm(vec[["y"]], mean = 3.02)))
}

# Define grid using lower and upper bounds and step widths
gridsearch(fn,
           lower = c(x = -10, y = -10),
           upper = c(x = 10, y = 10),
           step = c(x = 0.5, y = 0.5))
# Define grid using axes
gridsearch(fn,
           axes = list(x = (-10:10), y = (-10:10)))
# Custom grid, e.g. only the diagonal
grid <- data.frame(x = (-10:10), y = (-10:10))
gridsearch(fn,
           grid = grid)
# Diagnostics with progress bar
# Attention: Progress bar impedes performance!
\dontrun{
progressr::handlers(global = TRUE)
fn <- function(vec){
  Sys.sleep(0.1)
  return(c(result =
  (-1)*dnorm(vec[["x"]], mean = 3.9)*dnorm(vec[["y"]], mean = 3.02)))
}
gridsearch(fn,
           lower = c(x = -10, y = -10),
           upper = c(x = 10, y = 10),
           step = c(x = 0.5, y = 0.5),
           control = list(REPORT = 1))
}
}
