% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simann.R
\name{phiinv}
\alias{phiinv}
\title{Transform the unit interval to another interval}
\usage{
phiinv(x, lower, upper)
}
\arguments{
\item{x}{any real number}

\item{lower}{a real number, the lower bound of the interval}

\item{upper}{a real number, the upper bound of the interval}
}
\value{
a real number
}
\description{
A monotonously increasing bijection on the real line that maps the unit
interval \eqn{[0, 1]} to the interval \eqn{[\text{lower}, \text{upper}]}.
}
