% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simann.R
\name{phi}
\alias{phi}
\title{Transform an interval to the unit interval}
\usage{
phi(x, lower, upper)
}
\arguments{
\item{x}{any real number}

\item{lower}{a real number, the lower bound of the interval}

\item{upper}{a real number, the upper bound of the interval}
}
\value{
a real number
}
\description{
A monotonously increasing bijection on the real line that maps the interval
$\link{l, u}$ to the unit interval $\link{0, 1}$.
}
\examples{
phiinv(phi(7, 2, 3), 2, 3) == 7
}
