% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simann.R
\name{simann}
\alias{simann}
\title{Simulated Annealing}
\usage{
simann(par, fn, lower = NULL, upper = NULL, control)
}
\arguments{
\item{par}{numeric, start value for the algorithm}

\item{fn}{the function to be minimized}

\item{lower}{numeric, lower boundary of the parameter vector}

\item{upper}{numeric, upper boundary of the parameter vector}

\item{control}{a list of further control parameters for the algorithm}
}
\value{
a list containing \code{par}, the parameter combination of the optimum,
\code{value}, the optimal function value, and \code{trace}, a \code{data.frame} of all visited
parameter values.
}
\description{
This is almost a precise re-write of the simulated annealing
function \code{stats::optim(method = "SANN")} as implemented in the file
\code{r-source/src/appl/optim.c}. In particular, temperature is updated
step by step with respect to the formula
$$ T^{(i)} = T_{\text{start}} / \log(((i-1) \%/\% tmax)*tmax + exp(1))$$
and new candidate points are generated using a Gaussian kernel
$$ p^{(i)}\emph{j = p^{(i-1)} + \frac{T^{(i)}}{T}{\text{start}}} \cdot \varepsilon,$$
}
\details{
where $\varepsilon\sim\mathcal N(0, 1)$.

However, there are three main differences:
\itemize{
\item This function is currently implemented completely in R, for easier debugging.
\item This function returns a trace of all visited values as a \code{data.frame}.
\item This function allows for bounded parameter spaces by using the methodology
from chapter 6 of (Haario and Saksman 1991).
}

The adaption for the case of bounded parameter spaces is as follows:
W.l.o.g. assume that boundaries are $\link{0,1}^n$. Let $p$ the current point
and $p^{\text{try}}$ the candidate generated by the Gauss kernel. Suppose
that $i$-th component is not in boundary, i.e. $p^{\text{try}}_i\notin \link{0, 1}$.
Then division by 2 with remainder gives us $p^{\text{try}}_i\mod 2\in [0, 2)$,
when identifying $\mathbb R/2\mathbb Z\cong [0, 2)$ as sets. Then either
$p^{\text{try}}_i\mod 2\in \link{0, 1}$ or $2 - p^{\text{try}}_i \mod 2\in \link{0, 1}$.
Then, use this updated component of the candidate.
}
\section{Control parameters}{


The list \code{control} may contain the following objects:
\itemize{
\item \code{maxit}: the maximal number of temperature steps, defaults to \code{10000},
\item \code{temp}: the initial temperature, defaults to \code{10},
\item \code{tmax}: the maximal number of function evaluations at each
temperature step, defaults to \code{10},
\item \code{fnscale}: A scaling factor to be multiplied with the value of \code{fn} during
optimization. If it is negative, it turns the problem into a maximization
problem.
}
}

\examples{
fun <- function(x) {1 - dnorm(x, mean = 0, sd = 20)*10}
res <- simann(par = -10, fn = fun,
              control = list(maxit = 10000,
                             temp = 2000))
plot(res$trace$it, res$trace$fn)
plot(res$trace$it, res$trace$par1)
# With boundaries and named parameters
resbd <- simann(par = c(x = -1), fn = fun,
                lower = -2, upper = 2,
                control = list(maxit = 10000,
                               temp = 2000))
plot(resbd$trace$it, resbd$trace$fn)
plot(resbd$trace$it, resbd$trace$x)
}
\references{
Haario H, Saksman E. Simulated Annealing Process in General State Space. Advances in Applied Probability, Vol. 23, No. 4 (Dec., 1991), pp. 866-893, \link{https://doi.org/10.2307/1427681}.
}
